requirements.txt
Flask
PyYAML
numpy
opencv-python
pyrealsense2
